index: tests: split spec.lua per language
spec.lua had grown past 1.1k lines with every language inline. Each
language now lives in tests/lang/<lang>.lua, the case/idx/has/lacks
runner moved to tests/helpers.lua, and spec.lua is reduced to
alphabetized requires plus a final report() call.

The failures table is kept as an upvalue inside helpers.lua, so report()
still aggregates across all files and one broken case does not abort the
rest.
